requirements.txt
aiogram[fast]
asyncpg
google-generativeai
loguru
python-dotenv
pillow